requirements.txt
django
djangorestframework
drf_spectacular
mysqlclient
Pillow
